+++ MauiRPNCalculator/MauiRPNCalculator/App.xaml
