+++ MauiRPNCalculator/MauiRPNCalculator/App.xaml.cs
@@ -1,0 +1,12 @@
+﻿namespace MauiRPNCalculator
+{
+	public partial class App : Application
+	{
+		public App()
+		{
+			InitializeComponent();
+
+			MainPage = new AppShell();
+		}
+	}
+}
