+++ MauiRPNCalculator/MauiRPNCalculator/AppShell.xaml
