+++ MauiRPNCalculator/MauiRPNCalculator/AppShell.xaml.cs
