+++ MauiRPNCalculator/MauiRPNCalculator/MainPage.xaml.cs
@@ -1,0 +1,11 @@
+﻿namespace MauiRPNCalculator
+{
+	public partial class MainPage : ContentPage
+	{
+		int count = 0;
+		public MainPage()
+		{
+			InitializeComponent();
+		}
+	}
+}
