+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Android/AndroidManifest.xml
