+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Android/MainActivity.cs
