+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Android/MainApplication.cs
