+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/MacCatalyst/AppDelegate.cs
