+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/MacCatalyst/Entitlements.plist
