+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/MacCatalyst/Info.plist
