+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/MacCatalyst/Program.cs
