+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Tizen/Main.cs
