+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Tizen/tizen-manifest.xml
