+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Windows/App.xaml
