+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Windows/App.xaml.cs
