+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Windows/Package.appxmanifest
