+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/Windows/app.manifest
