+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/iOS/AppDelegate.cs
