+++ MauiRPNCalculator/MauiRPNCalculator/Platforms/iOS/Resources/PrivacyInfo.xcprivacy
