+++ MauiRPNCalculator/MauiRPNCalculator/Resources/Raw/AboutAssets.txt
