+++ MauiRPNCalculator/MauiRPNCalculator/Resources/Styles/Colors.xaml
