+++ MauiRPNCalculator/MauiRPNCalculator/Resources/Styles/Styles.xaml
