--- MauiRPNCalculator/Mvvm/Views/MainPage.xaml.cs
@@ -1,14 +1,0 @@
-﻿using MauiRPNCalculator.Mvvm.ViewModels;
-
-namespace MauiRPNCalculator
-{
-	public partial class MainPage : ContentPage
-	{
-		int count = 0;
-		public MainPage(MainPageViewModel vm)
-		{
-			InitializeComponent();
-			BindingContext = vm;	
-		}
-	}
-}
